Move Inditex accesstoken URL to env file, cache token using singleton
