+++ apikeys.example.env
@@ -1,2 +1,3 @@
-USER_INDITEX_API=
-SECRET_INDITEX_API=
+USER_INDITEX_API=<Replace by your OAuth2 Client ID>
+SECRET_INDITEX_API=<Replace by your OAuth2 Secret>
+ACCESTOKEN_URL_INDITEX_API=https://auth.inditex.com:443/openam/oauth2/itxid/itxidmp/access_token
