+++ clothing_compare/clothing_app/external_apis/singleton.py
@@ -1,0 +1,7 @@
+# Singleton class to ensure only one instance of a given class is created
+class Singleton(type):
+    _instances = {}
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
+        return cls._instances[cls]
