+++ compose.yml
@@ -6,7 +6,8 @@ services:
     volumes:
       - ./clothing_compare:/app
     env_file:
-      ".apikeys.env"
+      - ".apikeys.env"
+      - "models.env"
   frontend:
     build: 
       context: frontend
@@ -23,8 +24,10 @@ services:
     volumes:
       - ./model_data/ollama:/root/.ollama
     restart: unless-stopped
+    env_file:
+      - "models.env"
     post_start:
-      - command: ["sh", "-c", "ollama pull llama3.2:3b && ollama pull llava:7b"]
+      - command: ["sh", "-c", "ollama pull $${TEXT_MODEL} && ollama pull $${VISUAL_MODEL}"]
     # Enable GPU support
     # deploy:
     #       resources:
