+++ models.env
@@ -1,0 +1,2 @@
+TEXT_MODEL=llama3.2:3b
+VISUAL_MODEL=llava:7b
